move original code to new select model; adjust hjkl to display basic file system. create resources pkg
