+++ pkg/models/select/model.go
@@ -1,25 +1,26 @@
-package hjkl
+package select
 
 import (
-    	"fmt"
-        "log"
-        "os"
+	"fmt"
 
 	tea "github.com/charmbracelet/bubbletea"
-
-    
-    res "github.com/friedrichwilken/fd/pkg/resources"
 )
 
 type selection struct {
 	choices  []string         // items on the to-do list
 	cursor   int              // which to-do list item our cursor is pointing at
+	selected map[int]struct{} // which to-do items are selected
 }
 
 func New() selection{
 	return selection{
 		// Our to-do list is a grocery list
-		choices: dirContent(), 
+		choices: []string{"Buy carrots", "Buy celery", "Buy kohlrabi"},
+
+		// A map which indicates which choices are selected. We're using
+		// the  map like a mathematical set. The keys refer to the indexes
+		// of the `choices` slice, above.
+		selected: make(map[int]struct{}),
 	}
 }
 
@@ -56,8 +57,13 @@ func (m selection) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		// The "enter" key and the spacebar (a literal space) toggle
 		// the selected state for the item that the cursor is pointing at.
 		case "enter", " ":
-		    //todo move one dir down
-        }
+			_, ok := m.selected[m.cursor]
+			if ok {
+				delete(m.selected, m.cursor)
+			} else {
+				m.selected[m.cursor] = struct{}{}
+			}
+		}
 	}
 
 	// Return the updated model to the Bubble Tea runtime for processing.
@@ -79,13 +85,13 @@ func (m selection) View() string {
 		}
 
 		// Is this choice selected?
-		//checked := " " // not selected
-		//if _, ok := m.selected[i]; ok {
-		//	checked = "x" // selected!
-		//}
+		checked := " " // not selected
+		if _, ok := m.selected[i]; ok {
+			checked = "x" // selected!
+		}
 
 		// Render the row
-		s += fmt.Sprintf("%s %s\n", cursor, choice)
+		s += fmt.Sprintf("%s [%s] %s\n", cursor, checked, choice)
 	}
 
 	// The footer
@@ -93,30 +99,4 @@ func (m selection) View() string {
 
 	// Send the UI for rendering
 	return s
-}
-
-func dirContent() []string {    
-    entries, err := os.ReadDir("./")
-    if err != nil {
-        log.Fatal(err)
-    }
- 
-    names := []string{}
-    for _, e := range entries {
-        names = append(names, entryToString(e))
-    }
-    return names
-}
-
-func entryToString(e os.DirEntry) string {
-    // The retruned string will be either "<filesymbol> filename" or "<dirsymbol> dirname"
-    // Define if we use a file or dir symbol
-    s := ""
-    if e.IsDir() {
-        s = res.DirSymbol 
-    } else {
-        s = res.FileSymbol
-    }
-
-    return fmt.Sprintf("%s %s", s, e.Name())
 }
