+++ pkg/resources/symbols.go
@@ -1,0 +1,8 @@
+// Package resources contains resources. todo
+package resources
+
+//todo keep this as defaults but also allow config file.
+const (
+    DirSymbol = ""
+    FileSymbol = ""
+)
